Change ta additions test.
