+++ src/test/java/com/google/sps/servlets/AddClassTATest.java
@@ -1,0 +1,101 @@
+package com.google.sps.servlets;
+
+import static org.junit.Assert.assertTrue;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
+
+import com.google.appengine.api.datastore.DatastoreService;
+import com.google.appengine.api.datastore.DatastoreServiceFactory;
+import com.google.appengine.api.datastore.Entity;
+import com.google.appengine.api.datastore.PreparedQuery;
+import com.google.appengine.api.datastore.Query;
+import com.google.appengine.tools.development.testing.LocalDatastoreServiceTestConfig;
+import com.google.appengine.tools.development.testing.LocalServiceTestHelper;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.UserRecord;
+import java.util.Collections;
+import java.util.List;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
+
+@RunWith(MockitoJUnitRunner.class)
+public class AddClassTATest {
+  private final LocalServiceTestHelper helper =
+      new LocalServiceTestHelper(new LocalDatastoreServiceTestConfig());
+
+  private DatastoreService datastore;
+
+  @Mock FirebaseAuth authInstance;
+
+  @InjectMocks AddClassTA addTA;
+
+  @Mock HttpServletRequest httpRequest;
+
+  @Mock HttpServletResponse httpResponse;
+
+  @Before
+  public void setUp() {
+    helper.setUp();
+    datastore = DatastoreServiceFactory.getDatastoreService();
+  }
+
+  @After
+  public void tearDown() {
+    helper.tearDown();
+  }
+
+  @Test
+  // For a class with no current TAs, add one TA
+  public void addOneTAEmptyList() throws Exception {
+    Entity user = new Entity("User");
+
+    user.setProperty("userEmail", "test@google.com");
+    user.setProperty("registeredClasses", Collections.emptyList());
+    user.setProperty("taClasses", Collections.emptyList());
+    user.setProperty("ownedClasses", Collections.emptyList());
+
+    datastore.put(user);
+
+    // Create examples for the TA email and class code
+    when(httpRequest.getParameter("taEmail")).thenReturn("test@google.com");
+    when(httpRequest.getParameter("classCode")).thenReturn("exampleClassCode");
+
+    // Firebase needs to return the correct user ID for the TA
+    UserRecord userRecord = mock(UserRecord.class);
+    when(authInstance.getUserByEmail("test@google.com")).thenReturn(userRecord);
+
+    addTA.doPost(httpRequest, httpResponse);
+
+    Query query = new Query("User");
+    PreparedQuery results = datastore.prepare(query);
+
+    // Add the class key to the user's TA classes list
+    for (Entity entity : results.asIterable()) {
+      if (entity.getProperty("userEmail") == "test@google.com") {
+        List<String> taClasses = (List) entity.getProperty("taClasses");
+        assertTrue(taClasses.contains("exampleClassCode"));
+      }
+    }
+  }
+
+  @Test
+  // Throw an exception if class key isn't correct
+  public void keyUnavailable() throws Exception {
+
+    // Create examples for the TA email and class code
+    when(httpRequest.getParameter("taEmail")).thenReturn("test@google.com");
+    when(httpRequest.getParameter("classCode")).thenReturn("testClassCode");
+
+    addTA.doPost(httpRequest, httpResponse);
+
+    verify(httpResponse).sendError(HttpServletResponse.SC_BAD_REQUEST);
+  }
+}
